refactor: LoginViewController를 LoginController로 통합
