+++ app/src/main/java/com/techcourse/controller/LoginController.java
@@ -19,6 +19,16 @@ public class LoginController {
 
     private static final Logger log = LoggerFactory.getLogger(LoginController.class);
 
+    @RequestMapping(value = "/login/view", method = RequestMethod.GET)
+    public ModelAndView getLogin(HttpServletRequest req, HttpServletResponse res) throws Exception {
+        return UserSession.getUserFrom(req.getSession())
+                          .map(user -> {
+                              log.info("logged in {}", user.getAccount());
+                              return new ModelAndView(new JspView("redirect:/index.jsp"));
+                          })
+                          .orElse(new ModelAndView(new JspView("/login.jsp")));
+    }
+
     @RequestMapping(value = "/login", method = RequestMethod.POST)
     public ModelAndView login(HttpServletRequest req, HttpServletResponse res) {
         if (UserSession.isLoggedIn(req.getSession())) {
