--- app/src/main/java/com/techcourse/controller/LoginViewController.java
@@ -1,27 +1,0 @@
-package com.techcourse.controller;
-
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import nextstep.mvc.view.JspView;
-import nextstep.mvc.view.ModelAndView;
-import nextstep.web.annotation.Controller;
-import nextstep.web.annotation.RequestMapping;
-import nextstep.web.support.RequestMethod;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-@Controller
-public class LoginViewController {
-
-    private static final Logger log = LoggerFactory.getLogger(LoginViewController.class);
-
-    @RequestMapping(value = "/login/view", method = RequestMethod.GET)
-    public ModelAndView getLogin(HttpServletRequest req, HttpServletResponse res) throws Exception {
-        return UserSession.getUserFrom(req.getSession())
-                .map(user -> {
-                    log.info("logged in {}", user.getAccount());
-                    return new ModelAndView(new JspView("redirect:/index.jsp"));
-                })
-                .orElse(new ModelAndView(new JspView("/login.jsp")));
-    }
-}
